backend/requirements.txt
Flask==1.0.2
Flask-Cors==3.0.8
python-dateutil==2.8.1
psycopg2==2.8.4
passlib==1.7.2
PyJWT==1.7.1
flask-restx==0.1.0
marshmallow==3.3.0